fix signed-unsigned conversion

+++ RGFW.h
@@ -6997,7 +6997,7 @@ HICON RGFW_loadHandleImage(u8* src, RGFW_area a, BOOL icon) {
 
 	for (i32 y = 0; (u32)y < a.h; y++) {
 		for (i32 x = 0; (u32)x < a.w; x++) {
-			i32 index = (y * a.w + x) * 4;
+			u32 index = ((u32)y * a.w + (u32)x) * 4;
 			u8 alpha = src[index + 3];
 			if (alpha < 128) {
 				SetPixel(maskDC, x, y, RGB(255, 255, 255));
